Extracted interfaces for services. Annotated @Primary with default implementations
